cost-optimization/gke-keda/cloud-pubsub/requirements.txt
google-cloud-pubsub~=2.23.0
google-auth~=2.34.0
